@@ -82,25 +82,25 @@
   "homepage": "https://github.com/allusion-app/Allusion#readme",
   "devDependencies": {
     "@svgr/webpack": "^6.1.2",
-    "@types/chrome": "^0.0.86",
+    "@types/chrome": "^0.0.195",
     "@types/fs-extra": "^9.0.0",
     "@types/jest": "^27.4.0",
     "@types/offscreencanvas": "^2019.6.2",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^18.0.18",
+    "@types/react-dom": "^18.0.6",
     "@types/react-window": "^1.8.5",
     "@types/utif": "^3.0.1",
     "@types/uuid": "^8.3.0",
-    "@typescript-eslint/eslint-plugin": "^5.9.0",
-    "@typescript-eslint/parser": "^5.9.0",
+    "@typescript-eslint/eslint-plugin": "^5.36.2",
+    "@typescript-eslint/parser": "^5.36.2",
     "css-loader": "^6.5.1",
-    "electron": "18.0.0",
-    "electron-builder": "^22.14.13",
-    "eslint": "^8.6.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
+    "electron": "20.0.3",
+    "electron-builder": "^23.3.3",
+    "eslint": "^8.23.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.31.7",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "html-webpack-plugin": "^5.5.0",
     "jest": "^27.4.7",
     "mini-css-extract-plugin": "^2.4.6",
@@ -119,22 +119,22 @@
     "webpack-cli": "^4.9.1"
   },
   "dependencies": {
-    "@floating-ui/core": "^0.6.1",
-    "@floating-ui/react-dom": "^0.6.0",
+    "@floating-ui/core": "^1.0.1",
+    "@floating-ui/react-dom": "^1.0.0",
     "chokidar": "^3.5.2",
     "comlink": "^4.3.1",
     "dexie": "^3.2.2",
     "dexie-export-import": "^1.0.3",
     "electron-updater": "^5.0.0",
     "fs-extra": "^10.0.0",
-    "mobx": "^6.3.12",
-    "mobx-react-lite": "^3.2.3",
+    "mobx": "^6.6.2",
+    "mobx-react-lite": "^3.4.0",
     "node-exiftool": "^2.3.0",
     "node-stream-zip": "^1.15.0",
-    "react": "^17.0.2",
-    "react-colorful": "^5.5.1",
-    "react-dom": "^17.0.2",
-    "react-window": "^1.8.6",
+    "react": "^18.2.0",
+    "react-colorful": "^5.6.1",
+    "react-dom": "^18.2.0",
+    "react-window": "^1.8.7",
     "sourcemapped-stacktrace": "^1.1.11",
     "utif": "^3.1.0",
     "uuid": "^8.3.2",
